sender: minor fix
