--- go/src/socialapi/workers/email/sender1/event.go
@@ -1,24 +1,0 @@
-package sender
-
-type Event struct {
-	Name       string
-	User       *User
-	Body       *Body
-	Properties map[string]interface{}
-}
-
-type BodyType int
-
-type Body struct {
-	Type    BodyType
-	Content string
-}
-
-type User struct {
-	Email, Username string
-}
-
-const (
-	TextBodyType BodyType = iota
-	HtmlBodyType
-)
